ci(merge-to-dev): non fast forward merge requires configured git
Merging the release tag is rarely a fast forward of `dev` branch.
When the released is prepared, new commits contribution could be
merged in `dev`.

This requires a merge commit and git requires name and email for
authoring it.

Gitlab-ci does not export the git authoring variables:

- `$GIT_AUTHOR_NAME`
- `$GIT_AUTHOR_EMAIL`
- `$GIT_COMMITTER_NAME`
- `$GIT_COMMITTER_EMAIL`

We need to extract the information either from `$CI_COMMIT_AUTHOR`
containing `NAME <EMAIL>`, or from `$GITLAB_USER_NAME` and
`$GITLAB_USER_EMAIL`.

* .gitlab-ci.yml (merge-to-dev): extract authoring information from
  `$CI_COMMIT_AUTHOR` and export git variables.
  Add nice logging sections.